adds styling fixes and nname change to file
