--- src/components/NetflixSlider/Content.js
@@ -1,31 +1,0 @@
-import React from "react"
-import IconCross from "./../Icons/IconCross"
-import "./Content.css"
-
-const Content = ({ apartment, onClose }) => (
-  <div className="content">
-    <div className="content__background">
-      <div className="content__background__shadow" />
-      <div
-        className="content__background__image"
-        style={{ "background-image": `url(${apartment.image})` }}
-      />
-    </div>
-    <div className="content__area">
-      <div className="content__area__container">
-        <div className="content__title">{apartment.street}</div>
-        <div className="content__description">
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque
-          et euismod ligula. Morbi mattis pretium eros, ut mollis leo tempus
-          eget. Sed in dui ac ipsum feugiat ultricies. Phasellus vestibulum enim
-          quis quam congue, non fringilla orci placerat. Praesent sollicitudin
-        </div>
-      </div>
-      <button className="content__close" onClick={onClose}>
-        <IconCross />
-      </button>
-    </div>
-  </div>
-)
-
-export default Content
