--- src/components/NetflixSlider/Item.css
@@ -1,13 +1,0 @@
-.item {
-  flex: 0 0 19.7%;
-  transition: transform 300ms ease 100ms;
-  margin: 0 2px;
-  position: relative;
-  display: flex;
-
-  img {
-    height: 25%;
-    width: 25%;
-    vertical-align: top;
-  }
-}
