--- src/components/NetflixSlider/SliderWrapper.css
@@ -1,6 +1,0 @@
-.slider-wrapper {
-  padding: 40px 0;
-  overflow: hidden;
-  position: relative;
-  height: 200px;
-}
