+++ src/components/Slider/Content.css
@@ -1,0 +1,80 @@
+.content {
+	position: relative;
+	height: 37vw;
+	margin-top: -40px;
+}
+
+.content-background,
+.content-background-shadow,
+.content-background-image,
+.content-area {
+	position: absolute;
+	top: 0;
+	bottom: 0;
+}
+
+.content-background {
+	left: 0;
+	right: 0;
+}
+.content-shadow {
+	left: 0;
+	background: #000;
+	width: 30%;
+	z-index: 2;
+}
+.content-shadow:before {
+	content: "";
+	position: absolute;
+	z-index: 10;
+	background-image: linear-gradient(to right, #000, transparent);
+	top: 0;
+	bottom: 0;
+	left: 100%;
+	width: 275px;
+}
+
+.content-image {
+	right: 0;
+	width: 70%;
+	height: 100%;
+	background-position: center 10%;
+	background-size: cover;
+	z-index: 1;
+}
+
+.content-area {
+	left: 0;
+	right: 0;
+	height: 100%;
+	z-index: 3;
+}
+.content-area-container {
+	padding: 30px 70px;
+	color: wheat;
+}
+
+.content-title {
+	font-size: 45px;
+	color: #fff;
+	font-weight: 700;
+}
+
+.content-description {
+	font-size: 18px;
+	color: #999;
+	margin-top: 20px;
+	max-width: 500px;
+}
+
+.content-close {
+	color: #fff;
+	width: 40px;
+	height: 40px;
+	background: transparent;
+	outline: none;
+	border: none;
+	position: absolute;
+	top: 30px;
+	right: 30px;
+}
