+++ src/components/Slider/Item.js
@@ -1,0 +1,29 @@
+import React from "react"
+import cx from "classnames"
+import SliderContext from "./context"
+import ShowDetailsButton from "./ShowDetailsButton"
+import Mark from "./Mark"
+import "./Item.css"
+
+const Item = ({ apartment }) => (
+	<SliderContext.Consumer>
+		{({ onSelectSlide, currentSlide, elementRef }) => {
+			const isActive = currentSlide && currentSlide.id === apartment.id
+
+			return (
+				<div
+					ref={elementRef}
+					className={cx("item", {
+						"item-open": isActive,
+					})}
+				>
+					<img src={apartment.image} alt="" />
+					<ShowDetailsButton onClick={() => onSelectSlide(apartment)} />
+					{isActive && <Mark />}
+				</div>
+			)
+		}}
+	</SliderContext.Consumer>
+)
+
+export default Item
