+++ src/components/Slider/ShowDetailsButton.css
@@ -1,0 +1,19 @@
+.show-details-button {
+	position: absolute;
+	bottom: 0;
+	left: 0;
+	right: 0;
+	opacity: 0;
+	transition: opacity 300ms ease 100ms;
+	background: transparent;
+	border: 0;
+	outline: none;
+	width: 100%;
+
+	span {
+		display: block;
+		width: 12px;
+		margin: 0 auto;
+		color: white;
+	}
+}
