+++ src/components/Slider/SlideButton.js
