+++ src/components/Slider/Slider.css
@@ -1,0 +1,29 @@
+.slider {
+	display: flex;
+	position: relative;
+}
+
+.slider-container {
+	display: flex;
+	padding: 0 55px;
+	transition: transform 300ms ease 100ms;
+	z-index: 3;
+	width: 100%;
+	height: 200px;
+}
+
+.slider-container:not(.slider-open) .item:hover .show-details-button {
+	opacity: 1;
+}
+
+.slider-container:not(.slider-open) .item:hover {
+	transform: scale(1.5) !important;
+}
+
+.slider-container:not(.slider-open):hover .item {
+	transform: translateX(-25%);
+}
+
+.slider-container:not(.slider-open) .item:hover ~ .item {
+	transform: translateX(25%);
+}
