+++ src/components/Slider/SliderWrapper.css
@@ -1,0 +1,6 @@
+.slider-wrapper {
+	padding: 40px 0;
+	overflow: hidden;
+	position: relative;
+	height: 200px;
+}
