+++ src/components/Slider/SliderWrapper.js
