+++ src/components/Slider/context.js
