+++ src/components/Slider/useSizeElement.js
