+++ src/components/Slider/useSliding.js
